Split ChartWidgets
