+++ bin/VMSGUI
@@ -1,15 +1,10 @@
 #!/usr/bin/env python3.8
 
-import signal
-import sys
-
-from lsst.ts.cRIOpy.GUI import SALLog, SALComm, Application
-from lsst.ts.cRIOpy.VMS.ChartWidgets import *
-from lsst.ts.cRIOpy.VMS import ToolBar, StatusBar
-from lsst.ts.cRIOpy.VMS import Cache
+from lsst.ts.cRIOpy.GUI import SALLog, Application
+from lsst.ts.cRIOpy.VMS import Cache, ToolBar, StatusBar, BoxChartWidget, PSDWidget
 
 from PySide2.QtCore import Slot, Signal, QSettings, Qt
-from PySide2.QtWidgets import QMainWindow, QDockWidget
+from PySide2.QtWidgets import QMainWindow
 
 import astropy.units as u
 from asyncqt import asyncClose
