+++ python/lsst/ts/cRIOpy/VMS/Cache.py
@@ -19,7 +19,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-__all__ = ["VMSCache"]
+__all__ = ["Cache"]
 
 from .. import TimeCache
