+++ python/lsst/ts/cRIOpy/VMS/__init__.py
@@ -1,4 +1,6 @@
 from .Bars import StatusBar, ToolBar
+from .BoxChartWidget import BoxChartWidget
 from .Cache import Cache
-from .ChartWidgets import *
+from .ChartView import ChartView
+from .PSDWidget import PSDWidget
 from .TimeBoxChart import TimeBoxChart
